LPS-144374 Refactor ColorPicker component
