+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/js/index.js
@@ -12,4 +12,4 @@
  * details.
  */
 
-export {ColorPickerField} from '../page_editor/app/components/fragment-configuration-fields/ColorPickerField';
+export {ColorPicker} from '../page_editor/common/components/ColorPicker';
