+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -20,7 +20,6 @@
 @import 'Translation';
 @import 'ViewportSizeSelector';
 
-@import 'fragment-configuration-fields/ColorPickerField';
 @import 'fragment-configuration-fields/TargetCollectionsField';
 
 @import 'fragment-content/FragmentContent';
@@ -36,6 +35,7 @@
 @import '../../common/components/Collapse';
 @import '../../common/components/ColorPalette';
 @import '../../common/components/ColorPicker';
+@import '../../common/components/DropdownColorPicker';
 @import '../../common/components/Editor';
 @import '../../common/components/InlineConfirm';
 @import '../../common/components/ItemSelector';
