--- modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/fragment-configuration-fields/_ColorPickerField.scss
@@ -1,49 +1,0 @@
-html#{$cadmin-selector} {
-	.cadmin {
-		.page-editor__color-picker-field {
-			min-height: 32px;
-
-			.input-group
-				> .input-group-item
-				> .form-control.page-editor__color-picker-field__autocomplete__input {
-				border-radius: 0 4px 4px 0;
-				font-weight: 600;
-				transition: none;
-			}
-
-			.clay-color > .input-group-item > .input-group-text {
-				border-radius: 4px 0 0 4px;
-			}
-
-			&.has-warning {
-				.clay-color > .input-group-item > .input-group-text {
-					border-color: $cadmin-warning-l1;
-				}
-			}
-
-			&:not(.has-warning),
-			&.has-warning:not(:hover):not(:focus-within):not(.hovered) {
-				.input-group
-					> .input-group-item
-					> .form-control.page-editor__color-picker-field__autocomplete__input {
-					background-color: transparent;
-				}
-			}
-
-			&.hovered .page-editor__color-picker__selector {
-				background-color: $cadmin-gray-100;
-			}
-
-			&:not(:hover):not(:focus-within):not(.hovered) {
-				.page-editor__color-picker-field__action-button {
-					opacity: 0;
-				}
-
-				.clay-color > .input-group-item > .input-group-text,
-				.page-editor__color-picker-field__autocomplete__input {
-					border-color: transparent;
-				}
-			}
-		}
-	}
-}
