+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/common/components/_DropdownColorPicker.scss
@@ -1,0 +1,58 @@
+@import 'cadmin-variables';
+
+html#{$cadmin-selector} {
+	.cadmin {
+		.form-group-sm
+			.input-group
+			> .input-group-item
+			> .input-group-text.page-editor__dropdown-color-picker__input-group-text--rounded-left {
+			border-radius: 4px 0 0 4px;
+		}
+
+		.page-editor__dropdown-color-picker {
+			&__color-palette .sheet-subtitle {
+				border-color: $cadmin-secondary-l3;
+				border-width: 1px 0 0;
+				font-size: 0.75rem;
+			}
+
+			&__empty-result {
+				.c-empty-state-image {
+					max-width: 120px;
+
+					& + .c-empty-state-title {
+						font-size: 1rem;
+						margin-top: 32px;
+					}
+
+					& + .c-empty-state-text {
+						margin-top: 4px;
+					}
+				}
+			}
+
+			&__search-form {
+				&
+					+ .page-editor__dropdown-color-picker__color-palette
+					.sheet-subtitle {
+					border-width: 0;
+				}
+			}
+
+			&__selector {
+				&.btn {
+					height: 32px;
+					padding-left: 9px;
+					padding-right: 9px;
+				}
+
+				&-splotch {
+					border: 1px solid $cadmin-gray-300;
+					height: 20px;
+					margin-right: 18px;
+					min-width: 20px;
+				}
+			}
+		}
+	}
+}
